Fix bug that I've been trying to track down for months -
We would see weird VM corruption and such, with absolutely bizarre results,
including the PCPU_VIRTUAL mapping getting corrupted.  The TLB exception
refill code was doing everything right, but when it did a tlbp to check if the
entryhi was already in the TLB, it was using the previous value in k1 (entrylo1)
data to determine whether to do an indexed write or a random write, rather than
checking the TLB's index register.

The tlbp (TLB probe) instruction checks to see if any TLB entry matches the
current value of entryhi, and sets the TLB index register to the index in the
TLB at which the entry appears.  If it is negative (the high bit is set), then
the entry does not appear in the TLB.  When taking a TLB exception, one is
supposed to update the existing mapping, if there is one, so one sets up the
entryhi register and does a tlbp, and then uses tlbwi to update the entry which
corresponds to the TLB index register (the one we just found) if index is not
negative.  If it is negative, one is supposed to use tlbwr which writes to a
random TLB location (but not a wired location, ever.  A wired location was
getting overwritten, which is how it became clear something was doing a tlbwi
with a bogus TLB index register value.)