o) Reorganize pmap code so that cpu_pmap.c is the old page_map.c and:    o) paddr_t and vaddr_t are defined in cpu/types.h    o) MI PMAP interfaces are in cpu/pmap.h.    o) All things page table are in cpu/pte.h (including MD pmap functions.)    o) PAGE_SIZE is defined in cpu/page.h. o) Get rid of vm/types.h, whose only purpose is to forward-declare struct vm. o) Fix includes.
Now the badly-misnamed 'page' module no longer exists and all pmap and related
code is part of the main CPU module, which is probably as it ought to be.

And now code doesn't need to include VM and PMAP headers just to get at the
core address types and at PAGE_SIZE and friends.

I'm not perfectly happy with the new organization but as a better one becomes
apparent, thanks to the move to Subversion, I will reorganize appropriately.