o) Add an "mk-reconfig" top-level target which reruns mu-config in-place. o) Add braindead IPC data-passing structure and handling. o) Make NS handle registration requests. o) Make the ipc_service code handle registration responses from NS and the    proper formulation of a registration request. o) Modify IPC_HEADER_REPLY macro to yield a structure from a structure pointer    rather than modifying a structure in place. o) Implement a service directory using an (untested) BTREE. o) Annoyedly make the service name field of the NS structures "char" rather than    "unsigned char".
I have not tested a number of edge cases here and non-edge cases.  I've made a
few value judgments I'm not thrilled with and I suddenly think I need to have a
separate port for registering with NS so that random tasks can't do it.

I'm really not happy with how clumsy it feels to work with the IPC data stuff in
the microkernel.  Hopefully libraries can be developed to reduce the pain, at
very least in userland.  The microkernel can suffer under its sanity-checking,
but I'd hate to deal with this stuff when doing the servers.

This seems a fitting r1000, as it finally gives us IPC rich enough that we ought
to be able to do anything.  Next up, the 1000th change!  (cvs2svn took r1.)